[AIRLFOW-YYY] Move serve logs command to separate file
