+++ airflow/bin/cli.py
@@ -32,8 +32,8 @@ from tabulate import tabulate_formats
 from airflow import api, settings
 from airflow.cli.commands import (
     connection_command, dag_command, db_command, pool_command, role_command, rotate_fernet_key_command,
-    scheduler_command, sync_perm_command, task_command, user_command, variable_command, version_command,
-    webserver_command, worker_command,
+    scheduler_command, serve_logs_command, sync_perm_command, task_command, user_command, variable_command,
+    version_command, webserver_command, worker_command,
 )
 from airflow.configuration import conf
 from airflow.utils import cli as cli_utils
@@ -46,26 +46,6 @@ DAGS_FOLDER = settings.DAGS_FOLDER
 
 if "BUILDING_AIRFLOW_DOCS" in os.environ:
     DAGS_FOLDER = '[AIRFLOW_HOME]/dags'
-
-
-@cli_utils.action_logging
-def serve_logs(args):
-    """Serves logs generated by Worker"""
-    print("Starting flask")
-    import flask
-    flask_app = flask.Flask(__name__)
-
-    @flask_app.route('/log/<path:filename>')
-    def serve_logs(filename):  # pylint: disable=unused-variable, redefined-outer-name
-        log = os.path.expanduser(conf.get('core', 'BASE_LOG_FOLDER'))
-        return flask.send_from_directory(
-            log,
-            filename,
-            mimetype="application/json",
-            as_attachment=False)
-
-    worker_log_server_port = int(conf.get('celery', 'WORKER_LOG_SERVER_PORT'))
-    flask_app.run(host='0.0.0.0', port=worker_log_server_port)
 
 
 @cli_utils.action_logging
@@ -930,7 +910,7 @@ class CLIFactory:
             'args': ('principal', 'keytab', 'pid',
                      'daemon', 'stdout', 'stderr', 'log_file'),
         }, {
-            'func': serve_logs,
+            'func': serve_logs_command.serve_logs,
             'help': "Serve logs generate by worker",
             'args': tuple(),
         }, {
